@@ -46,83 +70,68 @@
   ],
   "license": "MIT",
   "devDependencies": {
-    "@types/base16": "^1.0.1",
-    "@types/enzyme": "^2.8.1",
-    "@types/isomorphic-fetch": "0.0.34",
-    "@types/jest": "^21.1.0",
-    "@types/node": "^8.0.2",
-    "@types/prop-types": "^15.5.1",
-    "@types/raven-js": "^3.10.0",
-    "@types/react": "^16.0.5",
-    "@types/react-addons-test-utils": "^0.14.19",
-    "@types/react-dom": "^15.5.1",
-    "@types/react-helmet": "^5.0.3",
-    "@types/react-intl": "^2.3.0",
-    "@types/react-redux": "^5.0.6",
-    "@types/redux": "^3.6.31",
-    "@types/redux-devtools": "^3.0.38",
-    "@types/redux-devtools-dock-monitor": "^1.1.31",
-    "@types/redux-devtools-log-monitor": "^1.0.31",
-    "@types/redux-form": "^7.0.5",
-    "@types/webpack": "^3.0.0",
-    "awesome-typescript-loader": "^3.0.0-beta.18",
-    "babel-core": "^6.22.1",
-    "babel-loader": "^7.0.0",
-    "babel-preset-es2015": "^6.22.0",
-    "cross-env": "^5.0.1",
-    "enzyme": "^2.9.1",
-    "file-loader": "^1.0.0",
-    "husky": "^0.14.0",
-    "jest": "^21.1.0",
-    "jest-enzyme": "^3.6.1",
-    "react-addons-test-utils": "^15.6.0",
-    "react-hot-loader": "^3.0.0-beta.7",
-    "react-test-renderer": "^15.5.4",
-    "rimraf": "^2.5.4",
-    "ts-jest": "^21.0.1",
-    "tslint": "^4.3.1",
-    "tslint-loader": "^3.3.0",
-    "tslint-microsoft-contrib": "^4.0.1",
-    "tslint-react": "^3.0.0",
-    "typescript": "^2.5.2",
-    "url-loader": "^0.5.7",
-    "validate-commit-msg": "^2.12.1",
-    "wallaby-webpack": "3.9.0",
-    "webpack": "^3.2.0",
-    "webpack-dev-middleware": "^1.11.0",
-    "webpack-dev-server": "^2.2.0",
-    "webpack-hot-middleware": "^2.15.0",
-    "webpack-manifest-plugin": "^1.1.0"
+    "@types/enzyme": "^3.1.13",
+    "@types/jest": "^23.3.1",
+    "@types/prop-types": "^15.5.5",
+    "@types/react": "^16.4.11",
+    "@types/react-dom": "^16.0.7",
+    "@types/react-helmet": "^5.0.7",
+    "@types/react-intl": "^2.3.9",
+    "@types/react-redux": "^6.0.6",
+    "@types/redux-form": "^7.4.6",
+    "awesome-typescript-loader": "^5.2.0",
+    "cross-env": "^5.2.0",
+    "enzyme": "^3.4.4",
+    "enzyme-adapter-react-16": "^1.2.0",
+    "file-loader": "^2.0.0",
+    "husky": "^1.0.0-rc.13",
+    "jest": "^23.5.0",
+    "jest-enzyme": "^6.0.4",
+    "react-hot-loader": "^4.3.5",
+    "rimraf": "^2.6.2",
+    "ts-jest": "^23.1.4",
+    "tslint": "^5.11.0",
+    "tslint-loader": "^3.6.0",
+    "tslint-microsoft-contrib": "^5.2.1",
+    "tslint-react": "^3.6.0",
+    "typescript": "^3.0.1",
+    "url-loader": "^1.1.1",
+    "validate-commit-msg": "^2.14.0",
+    "wallaby-webpack": "^3.9.11",
+    "webpack": "^4.17.1",
+    "webpack-cli": "^3.1.0",
+    "webpack-dev-middleware": "^3.2.0",
+    "webpack-hot-middleware": "^2.22.3",
+    "webpack-manifest-plugin": "^2.0.3"
   },
   "dependencies": {
-    "chalk": "^2.0.1",
-    "csstips": "^0.2.2",
-    "csx": "^8.2.1",
-    "es6-promise": "^4.0.5",
-    "express": "^4.14.0",
+    "chalk": "^2.4.1",
+    "csstips": "^0.3.0",
+    "csx": "^9.0.0",
+    "es6-promise": "^4.2.4",
+    "express": "^4.16.3",
     "isomorphic-fetch": "^2.2.1",
-    "prop-types": "^15.5.10",
-    "raven-for-redux": "^1.0.0",
-    "raven-js": "^3.16.0",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
-    "react-helmet": "^5.0.2",
-    "react-intl": "^2.3.0",
-    "react-redux": "^5.0.2",
-    "react-router5": "^5.0.2",
-    "redux": "^3.6.0",
-    "redux-form": "^7.0.4",
-    "redux-logger": "^3.0.0",
-    "redux-router5": "^5.0.0",
-    "redux-saga": "^0.15.4",
-    "router5": "^5.1.0",
-    "serve-favicon": "^2.3.2",
-    "typestyle": "^1.0.0"
+    "prop-types": "^15.6.2",
+    "raven-for-redux": "^1.3.1",
+    "raven-js": "^3.26.4",
+    "react": "^16.4.2",
+    "react-dom": "^16.4.2",
+    "react-helmet": "^5.2.0",
+    "react-intl": "^2.4.0",
+    "react-redux": "^5.0.7",
+    "react-router5": "^6.3.0",
+    "redux": "^4.0.0",
+    "redux-form": "^7.4.2",
+    "redux-logger": "^3.0.6",
+    "redux-router5": "^6.0.3",
+    "redux-saga": "^0.16.0",
+    "router5": "^6.4.2",
+    "serve-favicon": "^2.5.0",
+    "typestyle": "^2.0.1"
   },
   "jest": {
-    "mapCoverage": true,
     "transform": {
-      "^.+\\.tsx?$": "<rootDir>/node_modules/ts-jest/preprocessor.js"
+      "^.+\\.tsx?$": "ts-jest"
     },
     "cacheDirectory": "<rootDir>/.jest/cache",
     "collectCoverage": true,
