@@ -4,9 +4,11 @@ var config = {
   env: process.env.NODE_ENV || 'development',
   host: process.env.HOST || 'localhost',
   port: process.env.PORT || 8889,
-  karmaPort: 9876,
   ssr: true,
-  sentry: false,
+  sentry: {
+    dsn: '', // your sentry dsn here
+    options: {}
+  },
   // This part goes to React-Helmet for Head of our HTML
   app: {
     head: {
