@@ -2,10 +2,10 @@
  * Type declerations for global development variables
  */
 
+// tslint:disable:interface-name
+
 interface Window {
   // A hack for the Redux DevTools Chrome extension.
   __REDUX_DEVTOOLS_EXTENSION_COMPOSE__?: any;
   __INITIAL_STATE__?: any;
 }
-
-type Set<T> = any;
