@@ -12,6 +12,7 @@ fs.readdirSync('node_modules')
   });
 
 var config = {
+  mode: process.env.NODE_ENV === 'production' ? 'production' : 'development',
   externals: nodeModules,
   target: 'node',
 
@@ -30,27 +31,31 @@ var config = {
   },
 
   module: {
-    loaders: [
-      {
-        test: /\.(jpe?g|png|gif)$/i,
-        loader: 'url-loader?limit=1000&name=images/[hash].[ext]'
-      },
-      {
-        test: /\.jsx$/,
-        loader: 'babel-loader'
-      },
+    rules: [
       {
         test: /\.tsx?$/,
         loader: 'awesome-typescript-loader',
         exclude: /node_modules/
       },
       {
-        test: /\.woff(2)?(\?v=[0-9]\.[0-9]\.[0-9])?$/,
-        loader: "url-loader?limit=10000&mimetype=application/font-woff"
+        test: /\.eot(\?.*)?$/,
+        loader: 'file-loader?name=fonts/[hash].[ext]'
       },
       {
-        test: /\.(ttf|eot|svg)(\?v=[0-9]\.[0-9]\.[0-9])?$/,
-        loader: "file-loader"
+        test: /\.(woff|woff2)(\?.*)?$/,
+        loader: 'file-loader?name=fonts/[hash].[ext]'
+      },
+      {
+        test: /\.ttf(\?.*)?$/,
+        loader: 'url-loader?limit=10000&mimetype=application/octet-stream&name=fonts/[hash].[ext]'
+      },
+      {
+        test: /\.svg(\?.*)?$/,
+        loader: 'url-loader?limit=10000&mimetype=image/svg+xml&name=fonts/[hash].[ext]'
+      },
+      {
+        test: /\.(jpe?g|png|gif)$/i,
+        loader: 'url-loader?limit=10000&name=images/[hash].[ext]'
       }
     ]
   },
