@@ -4,7 +4,7 @@ import {connect} from "react-redux";
 import {actions} from "redux-router5";
 import {Router} from "router5";
 
-class Link extends React.Component<IDispatchToProps & IOwnProps, null> {
+class Link extends React.Component<IDispatchToProps & IOwnProps> {
   public static contextTypes: React.ValidationMap<any> = {
     router: object.isRequired
   };
@@ -38,6 +38,6 @@ interface IOwnProps {
   options?: any;
 }
 
-const connectedLink = connect<null, IDispatchToProps, IOwnProps>(null)(Link);
+const connectedLink = connect()(Link);
 
 export {connectedLink as Link, Link as UnconnectedLink};
