@@ -1,31 +1,33 @@
 import {shallow} from "enzyme";
 import * as React from "react";
 import {IAction, IState} from "../redux/modules/baseModule";
-import {ILanguage, SWITCH_LANGUAGE} from "../redux/modules/languageModule";
+import {CHANGE_LOCALE, IMeta as ISettingsMeta, ISettings} from "../redux/modules/settingsModule";
 import {mapStateToProps, UnconnectedAbout} from "./AboutPage";
 
 /* tslint:disable:no-empty jsx-no-lambda */
 describe("<AboutPage />", () => {
-  const language: IState<ILanguage> = {
-    payload: {
-      languageData: {greeting: "Hello!"},
+  const language: IState<ISettings, ISettingsMeta> = {
+    meta: {
       locale: "en-GB"
+    },
+    payload: {
+      translations: {greeting: "Hello!"}
     }
   };
 
   it("matches snapshot", () => {
-    const shallowComponent = shallow(<UnconnectedAbout dispatch={() => {}} locale=""/>);
+    const shallowComponent = shallow(<UnconnectedAbout dispatch={jest.fn()} locale=""/>);
     expect(shallowComponent).toMatchSnapshot();
   });
 
   it("maps state to props correctly", () => {
-    const props = mapStateToProps({language});
+    const props = mapStateToProps({settings: language});
     expect(props.locale).toEqual("en-GB");
   });
 
   it("calls switchLanguage() when button is clicked", () => {
     const spy = jest.spyOn(UnconnectedAbout.prototype, "switchLanguage");
-    const shallowComponent = shallow(<UnconnectedAbout dispatch={() => {}} locale=""/>);
+    const shallowComponent = shallow(<UnconnectedAbout dispatch={jest.fn()} locale=""/>);
 
     expect(shallowComponent.find("button")).toBeDefined();
     expect(spy).not.toHaveBeenCalled();
@@ -34,15 +36,14 @@ describe("<AboutPage />", () => {
   });
 
   describe("switchLanguage()", () => {
-    it("dispatches SWITCH_LANGUAGE action to en-GB", () => {
+    it("dispatches CHANGE_LOCALE action to en-GB", () => {
       const dispatch = jest.fn();
       const shallowComponent = shallow(<UnconnectedAbout dispatch={dispatch} locale="" />);
-      const expectedValue: IAction<ILanguage> = {
-        payload: {
-          languageData: {},
+      const expectedValue: IAction<ISettings, ISettingsMeta> = {
+        meta: {
           locale: "en-GB"
         },
-        type: SWITCH_LANGUAGE
+        type: CHANGE_LOCALE
       };
 
       expect(dispatch).not.toHaveBeenCalled();
@@ -50,15 +51,14 @@ describe("<AboutPage />", () => {
       expect(dispatch).toHaveBeenCalledWith(expectedValue);
     });
 
-    it("dispatches SWITCH_LANGUAGE action to de", () => {
+    it("dispatches CHANGE_LOCALE action to de", () => {
       const dispatch = jest.fn();
       const shallowComponent = shallow(<UnconnectedAbout dispatch={dispatch} locale="en-GB"/>);
-      const expectedValue: IAction<ILanguage> = {
-        payload: {
-          languageData: {},
+      const expectedValue: IAction<ISettings, ISettingsMeta> = {
+        meta: {
           locale: "de"
         },
-        type: SWITCH_LANGUAGE
+        type: CHANGE_LOCALE
       };
 
       expect(dispatch).not.toHaveBeenCalled();
