@@ -5,9 +5,9 @@ import {IStore} from "../redux/IStore";
 import {IDispatchToProps} from "../redux/modules/baseModule";
 import {decrement, increment} from "../redux/modules/counterModule";
 
-class CounterPage extends React.Component<IStateToProps & IDispatchToProps, null> {
-  constructor() {
-    super();
+class CounterPage extends React.Component<IStateToProps & IDispatchToProps> {
+  constructor(props: IStateToProps & IDispatchToProps) {
+    super(props);
     this.handleIncrement = this.handleIncrement.bind(this);
     this.handleDecrement = this.handleDecrement.bind(this);
   }
@@ -42,10 +42,10 @@ interface IStateToProps {
   count: number;
 }
 
-const mapStateToProps = (state: Partial<IStore>) => ({
+const mapStateToProps = (state: Pick<IStore, "counter">) => ({
   count: state.counter.payload.count
 });
 
-const connectedCounter = connect<IStateToProps, IDispatchToProps, null>(mapStateToProps)(CounterPage);
+const connectedCounter = connect<IStateToProps, IDispatchToProps>(mapStateToProps)(CounterPage);
 
-export {CounterPage as UnconnectedCounter, connectedCounter as CounterPage, mapStateToProps}
+export {CounterPage as UnconnectedCounter, connectedCounter as CounterPage, mapStateToProps};
