@@ -1,14 +1,16 @@
+import {color} from "csx";
 import * as React from "react";
 import {FormattedMessage} from "react-intl";
 import {style} from "typestyle";
 
 const styles = {
   home: style({
+    color: color("#125588").fade("80%").toString(),
     textAlign: "center"
   })
 };
 
-class HomePage extends React.Component<any, any> {
+class HomePage extends React.Component {
   public render(): JSX.Element {
     return (
       <div className={styles.home}>
@@ -19,4 +21,4 @@ class HomePage extends React.Component<any, any> {
   }
 }
 
-export {HomePage}
+export {HomePage};
