@@ -5,31 +5,43 @@ import {IStore} from "../redux/IStore";
 import {IDispatchToProps} from "../redux/modules/baseModule";
 import {loadStars} from "../redux/modules/starsModule";
 
-class StarsPage extends React.Component<IStateToProps & IDispatchToProps, null> {
-  public componentWillMount(): void {
-    if (this.props.stargazers_count === -1) {
+class StarsPage extends React.Component<IStateToProps & IDispatchToProps> {
+  constructor(props: IStateToProps & IDispatchToProps) {
+    super(props);
+
+    if (this.props.stargazersCount === -1) {
       this.props.dispatch(loadStars());
     }
   }
 
   public render(): JSX.Element {
-    const {isFetching, stargazers_count} = this.props;
     return (
       <div>
-        {isFetching ? <FormattedMessage id="stars.fetching" defaultMessage="Fetching Stars.." /> : stargazers_count}
+        {this.renderStars()}
       </div>
     );
+  }
+
+  private renderStars(): JSX.Element | string | number {
+    const {errorMessage, isFetching, stargazersCount} = this.props;
+    if (isFetching) {
+      return <FormattedMessage id="stars.fetching" defaultMessage="Fetching Stars.." />;
+    } else {
+      return errorMessage ? errorMessage : stargazersCount;
+    }
   }
 }
 
 interface IStateToProps {
   isFetching: boolean;
-  stargazers_count: number;
+  stargazersCount: number;
+  errorMessage?: string;
 }
 
-const mapStateToProps = (state: Partial<IStore>) => ({
+const mapStateToProps = (state: Pick<IStore, "stars">) => ({
+  errorMessage: state.stars.message,
   isFetching: state.stars.isFetching,
-  stargazers_count: state.stars.payload.stargazers_count
+  stargazersCount: state.stars.payload.stargazers_count
 });
-const connectedStars = connect<IStateToProps, IDispatchToProps, null>(mapStateToProps)(StarsPage);
-export {StarsPage as UnconnectedStars, connectedStars as StarsPage, mapStateToProps}
+const connectedStars = connect<IStateToProps, IDispatchToProps>(mapStateToProps)(StarsPage);
+export {StarsPage as UnconnectedStars, connectedStars as StarsPage, mapStateToProps};
