@@ -1,19 +1,14 @@
-import {State as IRouteState} from "router5";
+import {FormStateMap} from "redux-form";
+import {RouterState} from "redux-router5";
 import {IState} from "./modules/baseModule";
 import {ICounter} from "./modules/counterModule";
-import {ILanguage} from "./modules/languageModule";
+import {IMeta as ISettingsMeta, ISettings} from "./modules/settingsModule";
 import {IStars} from "./modules/starsModule";
-
-export interface IRouteReduxState {
-  route: IRouteState;
-  previousRoute?: IRouteState;
-  transitionRoute?: IRouteState;
-  transitionError?: string;
-}
 
 export interface IStore {
   counter: IState<ICounter>;
-  language: IState<ILanguage>;
-  router: IRouteReduxState;
+  form: FormStateMap;
+  settings: IState<ISettings, ISettingsMeta>;
+  router: RouterState;
   stars: IState<IStars>;
 }
