@@ -1,17 +1,19 @@
 import {Action} from "redux";
 
-export interface IAction<P> extends Action {
+export interface IAction<P, M = {}> extends Action {
   payload?: P;
   message?: string;
+  meta?: M;
 }
 
-export interface IState<P> {
+export interface IState<P, M = {}> {
   error?: boolean;
   isFetching?: boolean;
   message?: string;
+  meta?: M;
   payload: P;
 }
 
 export interface IDispatchToProps {
-  dispatch: (actionCreator: IAction<any>) => any;
+  dispatch: (actionCreator: IAction<any, any>) => any;
 }
