@@ -1,6 +1,5 @@
 import {IAction, IState} from "./baseModule";
-import * as counter from "./counterModule";
-import {DECREMENT, decrement, ICounter, INCREMENT, increment} from "./counterModule";
+import {counterReducer, DECREMENT, decrement, ICounter, INCREMENT, increment} from "./counterModule";
 
 describe("counterModule", () => {
   describe("reducer", () => {
@@ -13,20 +12,20 @@ describe("counterModule", () => {
           count: 0
         }
       };
-      expect(counter.counterReducer(undefined, {type: undefined})).toEqual(initialState);
+      expect(counterReducer(undefined, {type: undefined})).toEqual(initialState);
     });
 
     it("handles action of type INCREMENT", () => {
-      const action: IAction<ICounter> = {type: counter.INCREMENT};
-      expect(counter.counterReducer(state, action)).toEqual({
+      const action: IAction<ICounter> = {type: INCREMENT};
+      expect(counterReducer(state, action)).toEqual({
         isFetching: false,
         payload: {count: state.payload.count + 1}
       });
     });
 
     it("handles action of type DECREMENT when count is more than 1", () => {
-      const action: IAction<ICounter> = {type: counter.DECREMENT};
-      expect(counter.counterReducer(state, action)).toEqual({
+      const action: IAction<ICounter> = {type: DECREMENT};
+      expect(counterReducer(state, action)).toEqual({
         isFetching: false,
         payload: {count: state.payload.count - 1}
       });
@@ -34,15 +33,15 @@ describe("counterModule", () => {
 
     it("handles action of type DECREMENT when count is 1", () => {
       const stateLessThanOne: IState<ICounter> = {isFetching: false, payload: {count: 1}};
-      const action: IAction<ICounter> = {type: counter.DECREMENT};
-      expect(counter.counterReducer(stateLessThanOne, action)).toEqual({
+      const action: IAction<ICounter> = {type: DECREMENT};
+      expect(counterReducer(stateLessThanOne, action)).toEqual({
         isFetching: false,
         payload: {count: stateLessThanOne.payload.count - 1}
       });
     });
 
     it("handles actions with unknown type", () => {
-      expect(counter.counterReducer(state, {type: ""})).toEqual({
+      expect(counterReducer(state, {type: ""})).toEqual({
         isFetching: false,
         payload: {count: state.payload.count}
       });
