@@ -1,5 +1,6 @@
+import {promiseAction} from "../../helpers/promiseReducer";
 import {IAction, IState} from "./baseModule";
-import {IStars, LOAD_STARS, loadStars, requestType, starsReducer} from "./starsModule";
+import {IStars, LOAD_STARS, loadStars, starsReducer} from "./starsModule";
 
 describe("starsModule", () => {
   describe("reducer", () => {
@@ -19,19 +20,19 @@ describe("starsModule", () => {
       expect(starsReducer(stateBeforeAndAfter, action)).toEqual(stateBeforeAndAfter);
     });
 
-    it("handles action of type STARS_REQUEST", () => {
-      const action = {type: requestType.PENDING};
+    it("handles action of type LOAD_STARS_PENDING", () => {
+      const action = {type: promiseAction(LOAD_STARS).PENDING};
       const stateBefore: IState<IStars> = {isFetching: false, payload: null};
       const stateAfter: IState<IStars> = {isFetching: true, payload: null};
       expect(starsReducer(stateBefore, action)).toEqual(stateAfter);
     });
 
-    it("handles action of type STARS_SUCCESS", () => {
+    it("handles action of type LOAD_STARS_FULFILLED", () => {
       const action: IAction<IStars> = {
         payload: {
           stargazers_count: 99
         },
-        type: requestType.SUCCESS
+        type: promiseAction(LOAD_STARS).FULFILLED
       };
       const stateBefore: IState<IStars> = {isFetching: true, payload: null};
       const stateAfter: IState<IStars> = {
@@ -43,8 +44,8 @@ describe("starsModule", () => {
       expect(starsReducer(stateBefore, action)).toEqual(stateAfter);
     });
 
-    it("handles action of type STARS_FAILURE", () => {
-      const action: IAction<IStars> = {type: requestType.FAILURE, message: "error!"};
+    it("handles action of type LOAD_STARS_REJECTED", () => {
+      const action: IAction<IStars> = {type: promiseAction(LOAD_STARS).REJECTED, message: "error!"};
       const stateBefore: IState<IStars> = {isFetching: true, payload: null};
       const stateAfter: IState<IStars> = {error: true, isFetching: false, message: "error!", payload: null};
       expect(starsReducer(stateBefore, action)).toEqual(stateAfter);
