@@ -1,16 +1,16 @@
-import {combineReducers} from "redux";
+import {combineReducers, Reducer} from "redux";
 import {reducer as formReducer} from "redux-form";
 import {router5Reducer} from "redux-router5";
 import {IStore} from "./IStore";
 import {counterReducer} from "./modules/counterModule";
-import {languageReducer} from "./modules/languageModule";
+import {settingsReducer} from "./modules/settingsModule";
 import {starsReducer} from "./modules/starsModule";
 
-const rootReducer: Redux.Reducer<IStore> = combineReducers<IStore>({
+const rootReducer: Reducer<IStore> = combineReducers<IStore>({
   counter: counterReducer,
   form: formReducer,
-  language: languageReducer,
   router: router5Reducer,
+  settings: settingsReducer,
   stars: starsReducer
 });
 
