@@ -1,12 +1,19 @@
-import {Promise} from "es6-promise";
-import {ILanguage} from "../redux/modules/languageModule";
+import {ISettings} from "../redux/modules/settingsModule";
 import {IStars} from "../redux/modules/starsModule";
 
+// Don't forget to enable this in production!
+// tslint:disable:no-http-string
+
 export const dummyApi = {
-  getLanguageData: (payload: string): Promise<ILanguage> => {
-    return fetch(`http://localhost:8889/translation/${payload}`).then((res) => res.json());
+  getStars: (): Promise<IStars | {error: string}> => {
+    return fetch("https://api.github.com/repos/barbar/vortigern").then((res) => {
+      if (res.ok) {
+        return res.json();
+      }
+      return res.json().then((json) => ({error: json.message}));
+    });
   },
-  getStars: () : Promise<IStars> => {
-    return fetch("https://api.github.com/repos/barbar/vortigern").then((res) => res.json());
+  getTranslations: (payload: string): Promise<ISettings> => {
+    return fetch(`http://localhost:8889/translation/${payload}`).then((res) => res.json());
   }
 };
