@@ -9,7 +9,7 @@ import {RouterProvider} from "react-router5";
 import {App, Html} from "./app/containers";
 import {LanguageHelper} from "./app/helpers/LanguageHelper";
 import {configureStore} from "./app/redux/configureStore";
-import {ILanguage} from "./app/redux/modules/languageModule";
+import {ISettings} from "./app/redux/modules/settingsModule";
 import {configureRouter} from "./app/routes/configureRouter";
 import rootSaga from "./app/sagas/rootSaga";
 
@@ -23,7 +23,7 @@ const manifest = require("../build/manifest.json");
 const app = express();
 const translationHandler = (req, res) => {
   const languageHelper = new LanguageHelper(req.params.lang);
-  const lang: ILanguage = {languageData: languageHelper.getRequestLanguageData(), locale: languageHelper.getPreferredLanguage()};
+  const lang: ISettings = {translations: languageHelper.getRequestLanguageData()};
   res.json(lang);
 };
 
@@ -33,13 +33,9 @@ if (process.env.NODE_ENV !== "production") {
   const webpackCompiler = webpack(webpackConfig);
 
   app.use(require("webpack-dev-middleware")(webpackCompiler, {
-    historyApiFallback: true,
-    hot: true,
-    inline: true,
     lazy: false,
-    noInfo: true,
+    logLevel: "info",
     publicPath: webpackConfig.output.publicPath,
-    quiet: true,
     stats: {colors: true}
   }));
 
@@ -72,14 +68,20 @@ app.get("*", (req, res) => {
 
     const languageHelper = new LanguageHelper(req.headers["accept-language"]);
     const store = configureStore(router, {
-      language: {
-        payload: {
-          languageData: languageHelper.getRequestLanguageData(),
-          locale: languageHelper.getPreferredLanguage()
-        }
-      },
       router: {
-        route: routeState
+        previousRoute: null,
+        route: routeState,
+        transitionError: null,
+        transitionRoute: null
+      },
+      settings: {
+        meta: {
+          currency: "EUR",
+          locale: languageHelper.getPreferredLanguage()
+        },
+        payload: {
+          translations: languageHelper.getRequestLanguageData()
+        }
       }
     });
 
